Refactor tests that uses previously created retention
